Add files via upload
Here I worked on project .The objective of this project is to develop a
machine learning (ML) model that can predict the profit value of a
company based on its Administration Cost and Marketing Spend. We
used a dataset containing information from 50 companies, including
these variables and the corresponding profit earned.